src: Implement frame-rate independent logic across game entities
+ Add `frame_count_mod` to environment for scaling calculations relative
  to 60 FPS baseline.
+ Convert fixed frame-based updates in `satellite`, `menu`, `gameloop`,
  `explosion`, `floattext`, `tank`, and `player` to frame-rate
  independent logic.
+ Update movement, animations, and timers to ensure consistent behavior
  across varying frame rates.
+ Replace `int32_t` with `double` for position and velocity where
  sub-frame accuracy is required.
+ Adjust key handling logic to account for frame-rate scaled inputs.

+++ src/satellite.h
@@ -45,10 +45,10 @@ private:
 	 */
 
 	CBeam*   beam   = nullptr;
-	int32_t x      = 0;
-	int32_t y      = MENUHEIGHT + 5;
-	int32_t xv     = -2;
-	int32_t prev_x = 0;
+	double   x      = 0;
+	int32_t  y      = MENUHEIGHT + 5;
+	double   xv     = -2;
+	double   prev_x = 0;
 };
 
 #endif // ATANKS_SATELLITE_H_INCLUDED
